requirements.txt
openai>=1.0.0
twilio>=8.0.0
flask>=2.3.0
requests>=2.31.0
pillow>=10.0.0
python-dotenv>=1.0.0
pandas>=2.0.0
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0